requirements.txt
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.1.0
better_profanity==0.7.0
blockfrost-python==0.5.3
cbor2==5.4.6
certifi==2023.5.7
certvalidator==0.11.1
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cose==0.9.dev8
cryptography==41.0.3
Django==4.2.2
django-cors-headers==4.2.0
django-filter==23.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
docopt==0.6.2
ecdsa==0.18.0
ECPy==1.2.5
frozendict==2.3.8
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mnemonic==0.20
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==23.1
Pillow==9.5.0
pipreqs==0.4.13
pprintpp==0.4.0
psycopg2==2.9.6
pycardano==0.9.0
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typeguard==2.13.3
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
websocket-client==1.6.2
yarg==0.1.9